chore(ssi): refactor image resolver creation
The image resolver should not need to be passed through the system. What
we really need is the RC client, which we can then make a decision later
on how to use. This will allow us to refactor the image resolver without
neededing to modify how the cluster agent is boostrapped.