Allow for heroku variant build.
- Create agent "flavor" flags (none, fips, heroku)
- These can be used from the command line to install different pieces of the agent as needed.
- Use the distinction to install the correct set of OSS licenses.
- Fix license file for libpopt

Changing for heroku will invalidate the cache, but I can live with that
- it is a smaller build than the rest, an already a distinct pipeline
- for now, we are only using it for the packaging layers.  My plan is to adopt project files to be able to limit the scope of the flavor flag so that it will not reach outside of //packages/...

This is a precursor to #44033 and #44183, which add top level package targets for the distinct distribution artifacts.
At that time, the "select()"ion of all deps will happen from the top, and we won't need license generation in //compliance.