extract traceroute component (#43757)
### What does this PR do?

Extracts the traceroute runner into a component with two implementations: remote and local. Remote is used in non-system-probe processes, local used in system-probe.

### Motivation

Network path dynamic scheduling will run in system-probe when it is in direct send mode. This means we need the ability to trigger the traceroutes from within system-probe. This is the cleanest way to do that.

### Describe how you validated your changes

Existing tests.

### Additional Notes


Co-authored-by: bryce.kahle <bryce.kahle@datadoghq.com>