[EBPF-912] gpu: fix leak in kernel span pool (#43819)
### What does this PR do?

Fixes a leak in the kernel span pool that happened when no kernel launches matched the time filter. It fixes also a bug in the filter for memory allocations, discovered when making an equivalent test for memory spans.

### Motivation

Ensure all codepaths return spans to the pool correctly.

### Describe how you validated your changes

Added a test to reproduce the test.

### Additional Notes

Co-authored-by: guillermo.julian <guillermo.julian@datadoghq.com>