Set GIT_DEPTH to 0 for SQG (#44001)
### What does this PR do?
It sets GIT_DEPTH to 0 to ensure that we pull the entire commit history. It also fixes ownership of the configuration yaml file. Originally the ownership was set for Gitlab job definition and not threshold config.

### Motivation
We do see sometimes SQG are using wrong ancestor commit for size comparison ([example](https://github.com/DataDog/datadog-agent/pull/43843)).


Co-authored-by: joseph.gette <joseph.gette@datadoghq.com>