privileged logs: Add TOCTOU tests (#43868)
### What does this PR do?

There is code in the privileged logs module to prevent TOCTOU attacks
which change the file system between the evaluation of symbolic links
and the actual open of the file.

However, there are no tests cases for this since it's usually involves
timing.  To ensure that the handling is tested, add a hook point into
the validateAndOpenWithPrefix() function between the (known) times of
check and times of use to allow the tests to deterministally manipulate
the file system and simulate the attack.

### Motivation

Improve test coverage. Preparation for adding more allowed directories.


Co-authored-by: vincent.whitchurch <vincent.whitchurch@datadoghq.com>