privileged logs: Handle unsearchable directories (#43747)
### What does this PR do?

In any directory in the path of a configured log file does not have
executable (search) permissions, it does not get tailed even with
privileged logs enabled since a stat(2) call from the file launcher
fails and the logs agent never gets to the point where it calls into
system-probe.  Fix this by having that stat(2) call too fallback to
system-probe if privileged logs is enabled.

### Motivation

https://datadoghq.atlassian.net/browse/DSCVR-288

### Describe how you validated your changes

Unit and E2E tests have been modified to test this case.


Co-authored-by: vincent.whitchurch <vincent.whitchurch@datadoghq.com>