Replace all `filepath.Walk` by `filepath.WalkDir` (#43982)
### What does this PR do?

Replace all `filepath.Walk` by `filepath.WalkDir`.

### Motivation

On a staging cluster I used to profile the datadog agent, I noticed that the `filepath.Walk` called by the files logs launcher was consuming almost 4% of CPU.
<img width="3171" height="1136" alt="image" src="https://github.com/user-attachments/assets/2a406bf7-74c9-4504-a083-7f304fff03d0" />

The [Go documentation](https://pkg.go.dev/path/filepath#Walk) says that:
> Walk is less efficient than [WalkDir](https://pkg.go.dev/path/filepath#WalkDir), introduced in Go 1.16, which avoids calling `os.Lstat` on every visited file or directory.

And, when zooming in on the CPU profile, we can see that the said `Lstat` that can be saved by migrating from `filepath.Walk` to `filepath.WalkDir` represents almost 1.5% of CPU:
<img width="1808" height="1135" alt="image" src="https://github.com/user-attachments/assets/70ad0261-6ded-49bc-9aca-e2832a9d7086" />


### Describe how you validated your changes

No functional change is expected. Let the CI run the e2e tests.

### Additional Notes

I decided to migrate all the `filepath.Walk` even if they don’t appear on a hot path on profiles because there shouldn’t be any drawback of using the more modern API.
In some cases, we still need the `os.FileInfo` to check file permissions. In those cases, we call `d.Info()`, but only for the subset of files we’re interested in.

Co-authored-by: lenaic.huard <lenaic.huard@datadoghq.com>