chore(dda): fixup the option in new-e2e-tests.run to keep the stack
When running the dda command `new-e2e-tests.run` it shows an option to
keep the stack in end in order to re-run the tests `--keep_stack`.

This option does not work.

Keep the option, make it public by adding the option to the `--help` so
users can see it.

Make it so it sets the env variable `E2E_DEV_MODE` which is the current
option to keep the stack after running a test, this is the way.

Signed-off-by: Alexandre Lavigne <alexandre.lavigne@datadoghq.com>