[EBPF-956] gpu: fix detection of devices assigned to Docker containers (#43779)
### What does this PR do?

Fixes the detection of the devices assigned to Docker containers. Depending on how the agent is running, it might not have permissions to access the `/proc/\<pid>/environ` file and therefore it cannot know the value of the `NVIDIA_VISIBLE_DEVICES` variable. This PR adds a fallback where the check will query the Docker runtime by inspecting the container.

### Motivation

Fix for https://datadoghq.atlassian.net/browse/EBPF-956

### Describe how you validated your changes

Added e2e test code for this specific scenario, including a fix to ensure that the Docker socket is accessible in e2e tests. Part of these tests fulfill https://datadoghq.atlassian.net/browse/EBPF-952 and https://datadoghq.atlassian.net/browse/EBPF-771

### Additional Notes

This PR patches the GPU code only, so that the fix can be backported to 7.74.x with reduced risk. A future PR will ensure that this data is ingested in workloadmeta and used in a generic manner.

Co-authored-by: guillermo.julian <guillermo.julian@datadoghq.com>