[Backport 7.74.x] [EBPF-956] gpu: fix detection of devices assigned to Docker containers (PR #43779) (#44142)
(cherry picked from commit 12125c3acb5da121c6871d9b51ed143ba22f6055)


### What does this PR do?
Fixes the detection of the devices assigned to Docker containers. Depending on how the agent is running, it might not have permissions to access the `/proc/\<pid>/environ` file and therefore it cannot know the value of the `NVIDIA_VISIBLE_DEVICES` variable. This PR adds a fallback where the check will query the Docker runtime by inspecting the container.

### Motivation
Fix for https://datadoghq.atlassian.net/browse/EBPF-956

### Describe how you validated your changes
Added e2e test code for this specific scenario, including a fix to ensure that the Docker socket is accessible in e2e tests. Part of these tests fulfill https://datadoghq.atlassian.net/browse/EBPF-952 and https://datadoghq.atlassian.net/browse/EBPF-771

## Additional notes

This PR patches the GPU code only, so that the fix can be backported to 7.74.x with reduced risk. A future PR will ensure that this data is ingested in workloadmeta and used in a generic manner.



Co-authored-by: florent.clarret <florent.clarret@datadoghq.com>