[cluster-autoscaling] fix replica nodepool creation (#44117)
### What does this PR do?

- Fixes an issue where replica NodePool creation was not replicating the target NodePool spec
- Fixes an issue where a delay in LeaderElection resulted in the task not being requeued

More details:
Previously, a replica NodePool was being created based on the Target NodePool as designed, but it was being created only with the information coming from the backend - i.e. the labels, taints, and recommendations. Now the replica is copying the Spec of the Target NodePool, and then updating the `Spec.Requirements` with the recommendations.

For the second issue, a delay in LeaderElection would result in new NodePoolInternal objects in the store to never get reconciled. The NPI object would get added to the store correctly, but because `IsLeader()` is still `false` [here](https://github.com/DataDog/datadog-agent/blob/0bc5cd3d8ea3c210a61aa5c78d92a8834477cde0/pkg/clusteragent/autoscaling/cluster/controller.go#L110), the Process function exits and the item would not be requeued again (until a new payload is received from RC).

### Motivation

Realized the issue when doing an inventory check on all the features/expected behaviors.

### Describe how you validated your changes
Tested on a test EKS cluster linked to Staging

### Additional Notes


Co-authored-by: celene <celene@datadoghq.com>