[AGTMETRICS-393] Prevent DSD server from starting when data plane is enabled and handling DSD. (#43876)
### What does this PR do?

Ensures that the DogStatsD server in the Core Agent does not run when Agent Data Plane is enabled and serving DSD traffic.

### Motivation

Only one process can be handling DSD at a time, so this code is a strengthening of that invariant which we have been achieving, to date, by setting the appropriate environment variables for the respective processes. Doing so technically works, but is fragile and better handled directly in the Core Agent.

As part of doing this, we've introduced a new wrapper type for basic DSD configuration which currently only contains the logic around if DSD is enabled at all, and where it's enabled. This lets us centralize the logic as we depend on it in a few different places throughout the codebase. Our logic now considers the notion of DogStatsD being enabled _at all_ -- like, does the customer want DSD to be running, regardless of how we actually make that happen? -- and then separately can figure out _where_ it's enabled.

Additionally, we now skip displaying the `DogStatsD` section in the `agent status` output when DSD is disabled internally, as the metrics will always be zero and ADP will be pushing its status data to RAR and will effectively take over in terms of providing DSD status.

### Describe how you validated your changes

Built and ran the Agent locally, with a few different variations of config settings:

- all defaults: DSD runs in Core Agent, Core Agent-specific DSD section present in status output
- `use_dogstatsd` -> `false`: DSD does not run in Core Agent, Core Agent-specific DSD section not present in status output
- `data_plane.enabled`/`data_plane.dogstatsd.enabled` -> `true`: DSD does not run in Core Agent, Core Agent-specific DSD section not present in status output
- `use_dogstatsd` -> `false`, `data_plane.enabled`/`data_plane.dogstatsd.enabled` -> `true`: DSD does not run in Core Agent, Core Agent-specific DSD section not present in status output

Additionally, we've added new unit tests (and updated some existing ones) to ensure the logic is obeyed in terms of when DSD should and should not be running in the Core Agent.

### Additional Notes

We had to also fix up the logic of `SetConfig` on the `TestClient` helper, which is used to update configuration values in E2E tests. The previous logic only handled setting keys that went a level deep: `foo.bar`. As we needed to set `data_plane.dogstatsd.enabled`, this was insufficient... so we fixed the logic to allow for arbitrary nesting depth. We added a test for this while we were in there.


Co-authored-by: toby.lawrence <toby.lawrence@datadoghq.com>