[ABLD-278] Use bazel built libacl. (#44079)
Cutover from omnibus libacl to bazel

- Call libacl install directly from openscap.rb. When openscap is built with Bazel, we'll just depend on it as a dependency.
- Remove omnibus step for libacl
- Adjust build flags to match what configure/make would produce. 

### Testing:

- Compare 3 things
  - build with with configure/make,
  - pull libacl.so from an old CI image,
  - build with bazel and look at libacl.so

The sizes are close enough for comfort. They vary a bit because in the new build we are not double building some util functions, so we get a little space back.

Similar compare for the public header files.  Make sure that our new ones match previous agent builds.


Co-authored-by: tony.aiuto <tony.aiuto@datadoghq.com>