[CWS] do not follow inheritance of variables in status output (#44187)
### What does this PR do?

When displaying the value of each variables in status, we currently iterate on all PCEs to display process scoped variables. If a variable is inherited, we end up walking the whole ancestor tree, for each PCE. Quite wasteful. It's also not really needed, and creates a lot of noise in the status output.

This PR adds a boolean when querying the value, used to say "skip the inheritance part" and just get the direct value, and uses it for the status part.

### Motivation

### Describe how you validated your changes

### Additional Notes


Co-authored-by: paul.cacheux <paul.cacheux@datadoghq.com>