system-probe: Enable discovery by default if SP is enabled (#42675)
### What does this PR do?

Enable the discovery module if any other module is enabled in system-probe.
    
Currently, discovery is enabled automatically if USM is enabled, in order to
get process tags from the tracer metadata.  For the same reason, we will also
need to enable discovery if CNM is enabled.  Besides this, there is a stated
goal of enabling discovery by default in general in the future.  As a step
towards this, if any system-probe module is enabled, enable the discovery
module too, unless it is explicitly disabled.

### Describe how you validated your changes

Unit tests added.


Co-authored-by: vincent.whitchurch <vincent.whitchurch@datadoghq.com>