Mark gohai/network tests as flaky (#44282)
### What does this PR do?

Marks failing tests (such as in [this
job](https://gitlab.ddbuild.io/DataDog/datadog-agent/-/jobs/1298291692#L1278))
as flaky.

### Motivation

Mitigation for #incident-47105 to unblock the release.

The failures seem to be coming from interface configurations that are
not accounted for by neither the code nor the tests, concerning which
interfaces may or may not have a MAC address assigned.

### Describe how you validated your changes

### Additional Notes