discovery: Move normalization to core
Move service name normalization to the core agent instead of doing it in
system-probe.  This removes the need to import the normalization package
in SP (the core agent already imports it for other reasons) and also
removes the need to implement normalization in the upcoming light weight
SP for discovery.