Patch a generic placeholder in openssl libraries (#44219)
The problem is there is no official way to make openssl portable. Since we are triggering openssl's build system from inside of bazel we are getting sandbox paths injected into libraries and binaries created by openssl's build. The only current way to tackle it is to put a placeholder instead of the valid one (but openssl's build won't allow it) and then replace that placeholder with the expected installation path. So the workflow is following:
1. Build openssl
2. Patch libraries and replace sandbox paths with a PLACEHOLDER
3. Run @openssl//:install to put it to the expected place
4. Run //deps/openssl:fix_openssl_paths to replace PLACEHOLDER with the expected path

### What does this PR do?

### Motivation

### Describe how you validated your changes

### Additional Notes


Co-authored-by: spencer.gilbert <spencer.gilbert@datadoghq.com>