Fix deadlock in secret resolver
Secret would notify all subscribers when resolving a new config. This
behavior can create a deadlock depending on the subscriber's behavior.
We introduce a new parameter to silence notification based on the
caller's need.

We also introduce a way to remove origin so unscheduled checks can be
removed from the secret cache.