[EBPF-917] gpu: add initial CI job with GPUs (#43544)
### What does this PR do?

This PR enables tests on GPU runners, for improved integration testing and as an intermediate step between mocked unit tests and full e2e tests.

The new job only runs tests in two packages, specific for integration tests that will only run when there is a GPU available. These tests will be ignored in the rest of the test jobs. 

To avoid high cost and usage of GPU capacity, it's configured to not run in all `main` pipelines, only on scheduled ones. On PRs, it will only run if GPU-related code is changed.

This PR includes some basic tests that use real data from NVML, the test set will later be expanded.

### Motivation

Improve testing capabilities for GPUm.

### Describe how you validated your changes

CI passing.

### Additional Notes


Co-authored-by: guillermo.julian <guillermo.julian@datadoghq.com>