Bypass the gates if delta is 0 (#44393)
### What does this PR do?
It allows to bypass the static quality gates if reported delta is 0. Even this PR containing 0 code changes reports size change below 2 Kb (2015 bytes for one of the gates). Therefore, we apply rounding and treat changes below 2 Kb as no size change.

Also, we now upload reported delta to Datadog, that will allow us to build new dashboards for better visibility.

### Motivation
This should unblock people working on different aspects of the agent that don't touch agent's codebase, therefore, has no influence on the package size.

### Describe how you validated your changes
I lowered a limit for one of the gates to make it fail and see if SQG job is successful or not.


Co-authored-by: joseph.gette <joseph.gette@datadoghq.com>