dyninst/module: fix race with process removal and load
If a process went away while a program was being loaded, we could hit
a nil pointer panic. The issue was due to an invariant violation that
the runtime never returns a nil loaded program and a nil error. Code
has also been added to handle this bad behavior defensively, but the
bad behavior has been removed.