Remove schema source from nodetreemodel, improve performance (#44181)
### What does this PR do?

This PR removes the "schema" source layer from the config, along with the precomputed `allSettings` list.

The "schema" source layer was used for storing settings that don't have default values, because they were created using `SetKnown` or `BindEnv`, instead of `SetDefault` or `BindEnvAndSetDefault`. It would be accessed in three places: (1) the `Set` method in order to validate that we're not using `Set` to assign to an inner node, which some tests would do in order to create a subtree of settings at once, and (2) reading the yaml with `loadYamlInto` in order to know if the file layer did not match the shape of the default layer, and (3) to `computeAllSettings` which would be returned by `AllKeysLowercased`.

We can achieve the same above functionality by using the "default" layer instead, checking the `knownKeys` as well for settings without default values. This requires a slight modification to `knownKeys`, to store whether a key is for an inner node or a leaf.

Making this change greatly improves performance of `InitConfig` because we are no longer building an additional source layer, and also don't need to collect keys from this layer in `computeAllSettings`.

This PR is best reviewed one commit at a time.

### Motivation

Improve performance of nodetreemodel.

### Describe how you validated your changes

Unit tests cover this functionality. Some minor differences from Viper are also fixed by this change.

### Additional Notes


Co-authored-by: dustin.long <dustin.long@datadoghq.com>