[CWS] remove reference counting of process cache entries (#44225)
### What does this PR do?

We currently use a system of reference counting for process cache entries that is used to put back the PCE in the pool when the refcount drops to 0, but also to calls some cleanup callbacks in order to drop the associated variables.

This is mostly visible from the outside through the use of `Remain/Release` calls. This system is extremely difficult to use correctly, and has been the cause of multiple issues in the past.

This PR drops fully this reference counting system, in favor of letting the GC do it's job. PCE are still cached in the resolver, so they will be allocated per process and not per event, as is the case today.

In order to correctly drop the variables when a PCE is dropped, this PR uses [runtime.AddCleanup](https://pkg.go.dev/runtime#AddCleanup) to instructs the GC to drop the variables when "finalizing" the PCE.

### Motivation

### Describe how you validated your changes

### Additional Notes


Co-authored-by: paul.cacheux <paul.cacheux@datadoghq.com>