[ AGNTLOG-390 ] Extend core e2e test logs with a timestamp and the name of the test emitting the log (#43079)
### What does this PR do?
Currently, e2e tests will run a number of operations on remote machines via the test framework itself, rather than via the actual test code. This that the calling file and line number, both of whom are automatically prepended to logs emitted via the testing library's logging functionality, are not enough to identify which test actually emitted the log. This causes a traceability problem when multiple tests are being run in parallel.

There is already a framework in place for disambiguating logs, utilized in a high-value log emitting from host.go. This MR extends that slightly, targeting the most common files emitting logs in the test runs to shift over to a common logger.

Log lines under the new format:
```
    host.go:135: 03-12-2025 15:54:29 - TestLinuxVMFileTailingSuite - Reconnecting to host
    agent_commands.go:199: 03-12-2025 15:54:30 - TestLinuxVMFileTailingSuite - Waiting for the agent to be ready
    host.go:171: 03-12-2025 15:54:30 - TestLinuxVMFileTailingSuite - Executing command `sudo datadog-agent "status"`
=== RUN   TestLinuxVMFileTailingSuite/TestLinuxLogTailing
    suite.go:385: 03-12-2025 15:54:30 - TestLinuxVMFileTailingSuite/TestLinuxLogTailing - No change in provisioners, skipping environment update
```

### Motivation

### Describe how you validated your changes

### Additional Notes


Co-authored-by: ryan.hall <ryan.hall@datadoghq.com>