[ABLD-257] Migrate windows driver build to Bazel
- Create //packages/windows to hold the top level targets to build the MSI
- Call //packages/windows:install_deps from datadog-agent-dependencies.db
- Delete the datadog-windows-*-driver omnibus scripts.

This change assumes that WINDOWS_DDNPM_DRIVER and WINDOWS_DDPROCMON_DRIVER
should always be set during windows builds. That appears to be true
for omnibus builds, but that is the only thing we are changing right
now, so this should be safe.

The interesting question is if there is ever a need to build the windows
MSI without one or the other of those drivers.

Resolved in a slack thread:

- We don't need to worry about the case where the drivers are not assigned.
- There are multiple reasonable workflows, and the most common is to build the agent and then stomp with a custom driver. This PR is compatible with that.