fix(installer): Fix uninstall after remote upgrade (#44457)
### What does this PR do?
After a remote upgrade on Linux a few things happen:
- `/opt/datadog-agent` is removed to avoid having two installations on disk
- The new Agent lives under `/opt/datadog-packages` (notice the plural)
- Some files are created but not tracked by the deb / rpm, especially under `/opt/datadog-packages/run`

This means that calling the `prerm` from the installer under `/opt/datadog-agent` doesn't work anymore and files are left on disk. Additionally, the new files are not deleted but are owned by `dd-agent`, user which is deleted on a `purge`. These files become ownerless, and a subsequent reinstallation may create the user again with a different ID, causing ownership issues and breaking remote upgrades.

This PR fixes it by:
- Fixing the `/opt/datadog-packages/datadog-agent/stable` path to call the OCI installer if and only if it's not the same file as the deb installer (e.g. before the first remote upgrade)
  - This makes sure the Agent OCI doesn't remain installed as it is the case today
- Removing the additional files during the Agent's `prerm`
  - This fixes the aforementioned ownership issues on reinstallation

### Motivation
Fixing issues with uninstallation / reinstallation after a remote upgrade

### Describe how you validated your changes
Manual QA, new E2E test.

### Additional Notes
- Calling `remove` on an OCI does call `prerm` under the hood

Co-authored-by: baptiste.foy <baptiste.foy@datadoghq.com>