[CWS] Do not rate limit rules with 'fim' event type (#44482)
### What does this PR do?

Do not rate limit rules with 'fim' event type

### Motivation

The 'fim' event type makes uses of internal expanded rules for the different event types.
These rules have an internal ID that is not registered with the rate limiter, causing these rules
to be rate limited.

### Describe how you validated your changes

Create an agent rule with the "fim" event type, triggered the rule and checked that it was properly sent to the BE.

### Additional Notes

This is a regression introduced by https://github.com/DataDog/datadog-agent/pull/43553

Co-authored-by: sylvain.baubeau <sylvain.baubeau@datadoghq.com>
(cherry picked from commit d281b5fe556623c6aa8c4d8985567964eb57a472)