[AMCC-164]Allow tags to be removed from Distribution metrics (#44166)
### What does this PR do?

Allows tags to be removed from distribution metrics prior to aggregation.

### Motivation

Can be configured as:

```yaml

metric_tag_filterlist:
  metric_name:
    action: exclude
    tags:
    - tag1
    - tag2
 ```

Only distributions are supported. For the given metric with name `metric_name`,  if `tags_negated` is true then the provided tags will be removed from the metric. If `tags_negated` is false only the tags provided are kept.

This includes tags sent with the metric as well as tags added by the tagger (kubernetes tags etc...). Given the tagger tags are only added when resolving the context during aggregation, it is at this point that a wrapper over the TagAccumulator is created and it is that wrapper that is responsible for removing the tags.

The tagger maintains a cache of tags that are added to the metric. We can no longer always use this cache of tags since some of the tags may be removed. If a tag is removed from a tagger tag, a copy of these tags must be created. This will create more pressure on the allocator. Hopefully, the overhead will be minimal, but benchmarking is still needed to understand the impact. 


### Describe how you validated your changes

I have tested manually and added unit tests.


### Additional Notes


Co-authored-by: stephen.wakely <stephen.wakely@datadoghq.com>