[EBPF-995] gpu: fix metric type for errors metric (#44583)
Fixes the type being used for `gpu.errors` metrics, which should be emitted as gauge and not as a count, as we're emitting absolute values and not increases.

Ensure that metrics are presented correctly in the UI.

Unit tests passing.

Co-authored-by: guillermo.julian <guillermo.julian@datadoghq.com>
(cherry picked from commit dc7372385e6cc93698d1ca2addfed65b243836f0)