dyninst: add an id tying together chunks of a SymDB upload
Tag the different chunks (i.e. batches for packages) of a SymDB upload
with a common id so that the backend can recognize them as part of the
same upload. The backend needs to do that when trying to identify
duplicate uploads, and without this ID it has to use a heuristic based
on upload time.

Also tag the final chunk with a `final` bit so that the backend can
recognize the upload as complete.