APM: Prevent flakes in testServer for trace writer (APMSP-2499) (#44601)
### What does this PR do?
Prevent stray `|` characters from potentially panic-ing the TestServer implementation.

### Motivation
The test server uses incoming requests to setup expectations but incoming payloads can potentially include `|`, see testutil/span.go where we use the pipe char as a potential resource name for spans. To prevent a stray pipe from crashing the TestServer (and avoid completely re-writing how it handles expectations since this is just a test server) I modified the separator to be more unique. 

### Describe how you validated your changes
This flake seems to be _ultra_ rare (it only can happen when there is exactly 1 pipe character and when the value to the right happens to be a valid int, my best guess is that this can happen due to protobuf) I had a hard time reproducing it but given the stack trace this has to be it. I validated the test passes afterwards and didn't cause any regressions, this is all only test code too.

### Additional Notes


Co-authored-by: andrew.glaude <andrew.glaude@datadoghq.com>