Handle invalid file data like how viper does with GetViperCombine
Change how we handle merging config layers when there's invalid file data. If we need to merge nodes of different shape, we keep the higher layer's branch, unless it is a nil leaf. This means invalid file data is often kept in the merged root, but can be overriden by env vars, which improves compatibility with Viper.