Handle invalid file data like how viper does with GetViperCombine (#44605)
### What does this PR do?

Change how we handle merging config layers when there's invalid file data. If we need to merge nodes of different shape, we keep the higher layer's branch, unless it is a nil leaf. This means invalid file data is often kept in the merged root, but can be overridedn by env vars, which improves compatibility with Viper.

### Motivation

Compatibility with Viper

### Describe how you validated your changes

Unit tests

### Additional Notes


Co-authored-by: dustin.long <dustin.long@datadoghq.com>