[CWS] store program ID for cleanup in TC resolver (#44581)
### What does this PR do?

[Problematic profile](https://app.datadoghq.com/profiling/explorer?query=service%3Asystem-probe%20-runtime_os%3Awindows&agg_m=%40prof_core_cpu_cores&agg_m_source=base&agg_q=service&agg_q_source=base&agg_t=sum&attribute=rule_id&cols=timestamp%2Cservice%2C%40metrics.core_cpu_cores%2C%40metrics.core_alloc_bytes_per_sec%2Cversion%2Chost%2C%40metrics.go_lifetime_heap_bytes&event=AwAAAZtKt0af1LCkYQAAABhBWnRLdDBjQkFBRFVlYW5KVzhWZl93QUEAAAAkZjE5YjRhYjctZGYwOC00YTc3LTk0NDktYWM1NTU3MGVkYmM1AAOR4A&group_by=line&my_code=disabled&profileId=AZtKt0cBAADUeanJW8Vf_wAA&refresh_mode=paused&selection=%7B%22incl%22%3A%5B%22dynamic_linker_config_write%22%5D%7D&stream_sort=%40metrics.go_lifetime_heap_bytes%2Cdesc&top_n=100&top_o=top&viz=stream&x_missing=true&from_ts=1766483252315&to_ts=1766486852315&live=false)

If for some reason the probe has been unloaded from the outside of the TC resolver, the probe will set its internal ID to 0, and thus `.ID()` will return 0. This will cause the program to leak from the ebpf mapping, and make it impossible to cleanly removing the program entry. To fix this issue, this PR stores the program ID that was used in the mapping on the side of the probe itself, allowing to re-use the ID to clean it up, ensuring we don't leak any program behind in the mapping.

### Motivation

### Describe how you validated your changes

### Additional Notes


Co-authored-by: paul.cacheux <paul.cacheux@datadoghq.com>