[AGENTRUN-951] Shared library check e2e test: Don't remove the shlib after the test (#44589)
### What does this PR do?

This PR fixes the issue by not trying to remove the shlib. Having it removed would have cause issues to the Agent because the shlib check would still be scheduled.

Also. the e2e has been refactored to reduce code duplication between Linux and Windows

### Motivation

Fix the flakiness of the shlib check e2e test caused by the lock of the shlib during its removal.

### Describe how you validated your changes

e2e test in the CI.

### Additional Notes


Co-authored-by: maxime.chambre <maxime.chambre@datadoghq.com>