[CWS] Fix TestMoveMountRecursivePropagation (#44293)
### What does this PR do?

When inserting a new element to the mount resolver, sometimes it would detect a loop and stop looking for new children to be updated, which resulted in not all children being updated sometimes. Now it continues to look for other children when it fails.

Also, some fixes to the test. 


### Motivation

### Describe how you validated your changes

### Additional Notes


Co-authored-by: marcio.kovags <marcio.kovags@datadoghq.com>