Move nightly build workflow to modern conductor setting - datadog-agent part (#44643)
### What does this PR do?

This PR updates the `internal_kubernetes_deploy_experimental` job which triggers the deployment of Agent images built within the nightly pipeline, to use conductor as a deployment method instead of generating the CNAB workflow.

At the datadog-agent side this is mostly cleanup + passing the proper variables. The rest is in https://github.com/DataDog/k8s-datadog-agent-ops/pull/7046

### Motivation

Move the nightly deployments to conductor so we can block gitlab based deployment.

### Describe how you validated your changes

Tested locally in simulated pipelines runs. Full verification with the next nightly run.


Co-authored-by: kacper.murzyn <kacper.murzyn@datadoghq.com>